renamed the routes from `misc` to `footer`
