+++ portal/server/routes/index.js
@@ -3,7 +3,7 @@ const startRoutes = require('./start');
 const loginRoutes = require('./login');
 const dashboardRoutes = require('./dashboard');
 const contractRoutes = require('./contract');
-const miscRoutes = require('./misc');
+const footerRoutes = require('./footer.route');
 const adminRoutes = require('./admin');
 const userRoutes = require('./user');
 const feedbackRoutes = require('./feedback');
@@ -17,7 +17,7 @@ router.use('/', startRoutes);
 router.use('/', loginRoutes);
 router.use('/', dashboardRoutes);
 router.use('/', contractRoutes);
-router.use('/', miscRoutes);
+router.use('/', footerRoutes);
 router.use('/', adminRoutes);
 router.use('/', userRoutes);
 router.use('/', feedbackRoutes);
